+++ config.php
@@ -7,6 +7,6 @@ return [
     'class' => 'humhubContrib\auth\linkedin\Module',
     'namespace' => 'humhubContrib\auth\linkedin',
     'events' => [
-        [Collection::class, Collection::EVENT_AFTER_CLIENTS_SET, ['humhubContrib\auth\linkedin\Events', 'onAuthClientCollectionInit']]
+        [Collection::class, Collection::EVENT_AFTER_CLIENTS_SET, ['humhubContrib\auth\linkedin\Events', 'onAuthClientCollectionInit']],
     ],
 ];