+++ docs/CHANGELOG.md
@@ -1,9 +1,19 @@
 Changelog
 =========
 
-1.1.0 (Unreleased)
--------------------------
+1.2.0 (August 28, 2025)
+-----------------------
 - Fix: Update module resources path
+- Enh #9: Migration to Bootstrap 5 for HumHub 1.18
+
+1.1.1 (October 10, 2024)
+------------------------
+- Enh #7: Use PHP CS Fixer
+- Fix #6: Redirection to registration form fails (e.g. when a profile field is required) with version 2
+
+1.1.0 (June 7, 2024)
+--------------------
+- Fix #5: Fixed LinkedIn Auth Deprecations. Added `useV2` checkbox to switch beetween new and old auth clients
 
 1.0.1 (April 20, 2023)
 ----------------------